refactor: 테스트 코드 분리
테스트 클래스 분리
