+++ src/test/java/study/racingcar/RandomGeneratorTest.java
@@ -1,0 +1,26 @@
+package study.racingcar;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import study.racingcar.util.RandomGenerator;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
+public class RandomGeneratorTest {
+    private RandomGenerator randomGenerator;
+
+    @BeforeEach
+    void setup(){
+        randomGenerator = new RandomGenerator();
+    }
+
+    @Test
+    @DisplayName("랜덤 값 생성 범위가 0 ~ 9 맞는지 확인")
+    void getRandomNumber(){
+        assertThat(randomGenerator.getRandomNumber()).isLessThan(10);
+        assertThat(randomGenerator.getRandomNumber()).isLessThanOrEqualTo(9);
+        assertThat(randomGenerator.getRandomNumber()).isGreaterThan(-1);
+        assertThat(randomGenerator.getRandomNumber()).isGreaterThanOrEqualTo(0);
+    }
+}
